requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
cachetools==5.5.2
celery==5.5.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.2
defusedxml==0.7.1
dj-rest-auth==7.0.1
django==5.2
django-allauth==65.7.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
faker==37.1.0
fire==0.7.0
flower==2.0.1
google-auth==2.38.0
humanize==4.12.3
idna==3.10
kombu==5.5.4
markdown-it-py==3.0.0
mdurl==0.1.2
oauthlib==3.2.2
packaging==25.0
pip==25.1.1
prometheus-client==0.22.1
prompt-toolkit==3.0.51
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pygments==2.19.1
pyjwt==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2025.2
pyyaml==6.0.2
rav==0.0.9
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
termcolor==2.5.0
tornado==6.5.2
tzdata==2025.2
urllib3==2.3.0
uv==0.6.9
vine==5.1.0
wcwidth==0.2.13
stripe==11.6.0
